#572 added watch task to auto build and run tests. finished separation of modules into their own files for the most part
